[Rename] wrapper -> common 수정
